Add files via upload
rotary volume now reads the volume value before changing it. It ensures that it does not jump back to an old value if the volume has been changed by another application. radio button now emits a short "beep" to give the user a quick feedback that the command has been received.